backend/requirements.txt
# Core dependencies
fastapi>=0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
psycopg2-binary==2.9.9
pydantic>=2.0.0,<3.0.0
python-dotenv==1.0.0

# Database
asyncpg==0.29.0

# Utilities
python-dateutil==2.8.2

# AI/Agent dependencies
crewai>=0.28.8
crewai-tools>=0.1.6